RAG-Based-Profile-matching/requirements.txt
chromadb>=0.5.5
sentence-transformers>=3.0.1
transformers>=4.41.0
pypdf>=5.0.0
python-docx>=1.1.2
numpy>=1.26.0
